scripts/enrichment/requirements.txt
ruamel.yaml>=0.18.0
httpx>=0.27.0
